Switch to a prebuilt superuser v3.0.5 (39)
Super user now has Honeycomb code; and is using an API to provide
backwards compatibility.  However to build this compatibility layer
we need the Honeycomb SDK.

AS such I have used a Honeycomb compatible apktool to add:
android:sharedUserId="android.uid.superuser"
to the android manifest;

This will allow the block superuser updates feature to operate.

Otherwise this is the same binary as provided by ChainsDD